Break code into modular structures.
This adds a little overhead but is the first step toward making it
work on different platforms and with unit tests. Mainly abstracting
the ADC, the control module and the serial output.
The SampleBuffer should be replaceable as I think I'll need to move to
multiple buffers at some point - the original could result in serving
up ghost data if triggered too soon. 'Fixing' that has meant the output
can be variable length right now.

Also included the beginings of unit tests (just POC at this point) and
I'm not overly happy with it.

This is rapidly going in a different direction from the original project
so it's renamed to stackfull/scope to avoid awkward version conflicts
with the real Girino. It still contains much of the original code so the
copyright remains in place for those files and the licence remains GPL3.